feat(fin): endpoints de FX rates (FIN-011)\n\n- POST /fin/fx-rates: upsert por (date, base, quote); 201 criado, 200 atualizado\n- GET /fin/fx-rates?base=&quote=&from=&to=: lista ordenada, valida base!=quote e range\n- Schema: valida escala Decimal(<=10) e base!=quote\n- Model: rate_value como Decimal (Numeric(18,10))

+++ app/models/finance/fx_rate.py
@@ -1,6 +1,7 @@
 from sqlalchemy import Date, Numeric, String, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 import datetime as dt
+from decimal import Decimal
 
 from app.models.base import Base, TimestampMixin
 
@@ -12,9 +13,8 @@ class FxRate(TimestampMixin, Base):
     date: Mapped[dt.date] = mapped_column(Date, index=True)
     base: Mapped[str] = mapped_column(String(3), index=True)
     quote: Mapped[str] = mapped_column(String(3), index=True)
-    rate_value: Mapped[float] = mapped_column(Numeric(18, 10))
+    rate_value: Mapped[Decimal] = mapped_column(Numeric(18, 10))
 
     __table_args__ = (
         UniqueConstraint("date", "base", "quote", name="uq_fx_rates_date_base_quote"),
     )
-
